move model container from distributions.py to models.py
